add box type to app arguments
adds a box type to the app arguments in the margareth enum in `src/main.rs`. the `warn` attribute was also added to the enum to suppress a clippy warning.

the `app` variant was updated to accept a `box<app::args>` instead of `app::args`. the `run` function was also updated to dereference `app` when calling `app::run`.